+++ changes.md
@@ -4,3 +4,7 @@ Change
 * MyPlugin: added a PluginClass which holds the Widgets tree and selection
 * TreeWidget: changed the param type from **..check_box_state_changed from bool to int**
 * OverviewPlugin: added missing method **on_range_combo_box_index_changed(index)**
+
+## Configuration files
+* Now uses seuid where possible
+=> specifications now take **seuid**: {} instead of **type: name**: {} (see arni_processing/resources/testconfig.yaml)